parallel_enum v0.1.0 ParallelEnum

Some parallel processing function inspired by https://gist.github.com/padde/b350dbc81f458898f08417d804e793aa

Functions

map(list, fun, max \\ 1)

map(list | map, function, integer) :: list

Map all element of a list and execute the given function on it in parallel. If the function raise an error during one process still get going & return the error in the relevant result index as: > {:error, err}

map!(list, fun, max \\ 1)

map!(list | map, function, integer) :: list

Map all element of a list and execute the given function on it in parallel. Error are propagated to the main thread